refactor: isolate weapon drop generation
